removed a warning

+++ Server/worldSaver/worldSaver.cpp
@@ -53,8 +53,8 @@ void WorldSaver::save() {
         output.insert(output.end(), i.first.begin(), i.first.end());
         output.push_back(0);
         
-        output.insert(output.end(), {0, 0, 0, 0});
-        unsigned int size = i.second.size();
+        output.insert(output.end(), {0, 0, 0, 0}); 
+        unsigned int size = (unsigned int)i.second.size();
         memcpy(&output[output.size() - 4], &size, sizeof(unsigned int));
         
         output.insert(output.end(), i.second.begin(), i.second.end());
